strands_agents/service/requirements.txt
fastapi
uvicorn
pydantic
strands-agents[openai] @ git+https://github.com/strands-agents/sdk-python.git@main
strands-agents-tools
boto3
botocore
ollama
python-dotenv
pyyaml
